Remove ucar.nc2.dt package. Remove ui dependencies on dt. Add GridCoordinateSystem.isZPositive(); mostly a kludge for now.
