--- cdm-core/src/main/java/ucar/nc2/dt/grid/GridDataset.java
@@ -1,525 +1,0 @@
-/*
- * Copyright (c) 1998-2018 John Caron and University Corporation for Atmospheric Research/Unidata
- * See LICENSE for license information.
- */
-package ucar.nc2.dt.grid;
-
-import java.io.IOException;
-import java.util.ArrayList;
-import java.util.Formatter;
-import java.util.HashMap;
-import java.util.List;
-import java.util.Map;
-import java.util.Set;
-import ucar.nc2.Attribute;
-import ucar.nc2.AttributeContainer;
-import ucar.nc2.NetcdfFile;
-import ucar.nc2.Variable;
-import ucar.nc2.VariableSimpleIF;
-import ucar.nc2.constants.CDM;
-import ucar.nc2.constants.FeatureType;
-import ucar.nc2.dataset.*;
-import ucar.nc2.dataset.NetcdfDataset.Enhance;
-import ucar.nc2.dt.GridCoordSystem;
-import ucar.nc2.dt.GridDatatype;
-import ucar.nc2.ft.FeatureDataset;
-import ucar.nc2.calendar.CalendarDate;
-import ucar.nc2.calendar.CalendarDateRange;
-import ucar.nc2.internal.cache.FileCacheIF;
-import ucar.unidata.geoloc.LatLonRect;
-import ucar.unidata.geoloc.ProjectionRect;
-
-/**
- * Make a NetcdfDataset into a collection of GeoGrids with Georeferencing coordinate systems.
- * <p/>
- * <p/>
- * A variable will be made into a GeoGrid if it has a Georeferencing coordinate system,
- * using GridCoordSys.isGridCoordSys(), and it has no extra dimensions, ie
- * GridCoordSys.isComplete( var) is true.
- * If it has multiple Georeferencing coordinate systems, any one that is a product set will be given preference.
- * <p/>
- * Example:
- * <p/>
- * 
- * <pre>
- * GridDataset gridDs = GridDataset.open(uriString);
- * List grids = gridDs.getGrids();
- * for (int i = 0; i &lt; grids.size(); i++) {
- *   GeoGrid grid = (Geogrid) grids.get(i);
- * }
- * </pre>
- *
- * @author caron
- */
-
-public class GridDataset implements ucar.nc2.dt.GridDataset, FeatureDataset {
-  private NetcdfDataset ncd;
-  private final ArrayList<GeoGrid> grids = new ArrayList<>();
-  private final Map<String, Gridset> gridsetHash = new HashMap<>();
-
-  /**
-   * Open a netcdf dataset, using NetcdfDataset.defaultEnhanceMode plus CoordSystems
-   * and turn into a GridDataset.
-   *
-   * @param location netcdf dataset to open, using NetcdfDataset.acquireDataset().
-   * @return GridDataset
-   * @throws java.io.IOException on read error
-   */
-  public static GridDataset open(String location) throws java.io.IOException {
-    return open(location, NetcdfDataset.getDefaultEnhanceMode());
-  }
-
-  /**
-   * Open a netcdf dataset, using NetcdfDataset.defaultEnhanceMode plus CoordSystems
-   * and turn into a GridDataset.
-   *
-   * @param location netcdf dataset to open, using NetcdfDataset.acquireDataset().
-   * @param enhanceMode open netcdf dataset with this enhanceMode
-   * @return GridDataset
-   * @throws java.io.IOException on read error
-   */
-  public static GridDataset open(String location, Set<NetcdfDataset.Enhance> enhanceMode) throws java.io.IOException {
-    NetcdfDataset ds = ucar.nc2.dataset.NetcdfDatasets.acquireDataset(null, DatasetUrl.findDatasetUrl(location),
-        enhanceMode, -1, null, null);
-    return new GridDataset(ds, null);
-  }
-
-  /**
-   * Create a GridDataset from a NetcdfDataset.
-   *
-   * @param ncd underlying NetcdfDataset, will do Enhance.CoordSystems if not already done.
-   * @throws java.io.IOException on read error
-   */
-  public GridDataset(NetcdfDataset ncd) throws IOException {
-    this(ncd, null);
-  }
-
-  /**
-   * Create a GridDataset from a NetcdfDataset.
-   *
-   * @param ncd underlying NetcdfDataset, will do Enhance.CoordSystems if not already done.
-   * @param parseInfo put parse info here, may be null
-   * @throws java.io.IOException on read error
-   */
-  public GridDataset(NetcdfDataset ncd, Formatter parseInfo) throws IOException {
-    Set<Enhance> enhance = ncd.getEnhanceMode();
-    if (enhance == null || !enhance.contains(NetcdfDataset.Enhance.CoordSystems)) {
-      enhance = NetcdfDataset.getDefaultEnhanceMode();
-      this.ncd = NetcdfDatasets.enhance(ncd, enhance, null);
-    } else {
-      this.ncd = ncd;
-    }
-
-    // look for geoGrids
-    if (parseInfo != null)
-      parseInfo.format("GridDataset look for GeoGrids%n");
-    List<Variable> vars = ncd.getVariables();
-    for (Variable var : vars) {
-      VariableEnhanced varDS = (VariableEnhanced) var;
-      constructCoordinateSystems(ncd, varDS, parseInfo);
-    }
-  }
-
-  private void constructCoordinateSystems(NetcdfDataset ds, VariableEnhanced v, Formatter parseInfo) {
-    if (v instanceof StructureDS) {
-      StructureDS s = (StructureDS) v;
-      List<Variable> members = s.getVariables();
-      for (Variable nested : members) {
-        // LOOK flatten here ??
-        constructCoordinateSystems(ds, (VariableEnhanced) nested, parseInfo);
-      }
-    } else {
-      // see if it has a GridCS
-      // LOOK: should add geogrid it multiple times if there are multiple geoCS ??
-      GridCoordSys gcs = null;
-      List<CoordinateSystem> csys = v.getCoordinateSystems();
-      for (CoordinateSystem cs : csys) {
-        GridCoordSys gcsTry = GridCoordSys.makeGridCoordSys(parseInfo, cs, v);
-        if (gcsTry != null) {
-          gcs = gcsTry;
-          if (gcsTry.isProductSet())
-            break;
-        }
-      }
-
-      if (gcs != null)
-        addGeoGrid((VariableDS) v, gcs, parseInfo);
-    }
-
-  }
-
-  private LatLonRect llbbMax;
-  private CalendarDateRange dateRangeMax;
-  private ProjectionRect projBB;
-
-  private void makeRanges() {
-    LatLonRect.Builder llbbBuilder = null;
-    ProjectionRect.Builder projBBbuilder = null;
-    for (ucar.nc2.dt.GridDataset.Gridset gset : getGridsets()) {
-      GridCoordSystem gcs = gset.getGeoCoordSystem();
-
-      ProjectionRect bb = gcs.getBoundingBox();
-      if (projBBbuilder == null)
-        projBBbuilder = bb.toBuilder();
-      else
-        projBBbuilder.add(bb);
-
-      LatLonRect llbb = gcs.getLatLonBoundingBox();
-      if (llbbBuilder == null) {
-        llbbBuilder = llbb.toBuilder();
-      } else {
-        llbbBuilder = llbbBuilder.extend(llbb);
-      }
-
-      CalendarDateRange dateRange = gcs.getCalendarDateRange();
-      if (dateRange != null) {
-        if (dateRangeMax == null)
-          dateRangeMax = dateRange;
-        else
-          dateRangeMax = dateRangeMax.extend(dateRange);
-      }
-    }
-
-    if (llbbBuilder != null) {
-      llbbMax = llbbBuilder.build();
-    }
-
-    projBB = projBBbuilder.build();
-
-  }
-
-  // stuff to satisfy ucar.nc2.dt.TypedDataset
-  public String getTitle() {
-    String title = ncd.getTitle();
-    if (title == null)
-      title = ncd.getRootGroup().findAttributeString(CDM.TITLE, null);
-    if (title == null)
-      title = getName();
-    return title;
-  }
-
-  public String getDescription() {
-    String desc = ncd.getRootGroup().findAttributeString("description", null);
-    if (desc == null)
-      desc = ncd.getRootGroup().findAttributeString(CDM.HISTORY, null);
-    return (desc == null) ? getName() : desc;
-  }
-
-  public String getLocation() {
-    return (ncd != null) ? ncd.getLocation() : "";
-  }
-
-  public CalendarDateRange getCalendarDateRange() {
-    if (dateRangeMax == null)
-      makeRanges();
-    return dateRangeMax;
-  }
-
-  public CalendarDate getCalendarDateStart() {
-    if (dateRangeMax == null)
-      makeRanges();
-    return (dateRangeMax == null) ? null : dateRangeMax.getStart();
-  }
-
-  public CalendarDate getCalendarDateEnd() {
-    if (dateRangeMax == null)
-      makeRanges();
-    return (dateRangeMax == null) ? null : dateRangeMax.getEnd();
-  }
-
-  public LatLonRect getBoundingBox() {
-    if (llbbMax == null)
-      makeRanges();
-    return llbbMax;
-  }
-
-  public ProjectionRect getProjBoundingBox() {
-    if (llbbMax == null)
-      makeRanges();
-    return projBB;
-  }
-
-  public void calcBounds() {
-    // not needed
-  }
-
-  @Override
-  public AttributeContainer attributes() {
-    return ncd.getRootGroup().attributes();
-  }
-
-  @Override
-  public List<VariableSimpleIF> getDataVariables() {
-    List<VariableSimpleIF> result = new ArrayList<>(grids.size());
-    for (GridDatatype grid : getGrids()) {
-      if (grid.getVariable() != null) // LOOK could make Adaptor if no variable
-        result.add(grid.getVariable());
-    }
-    return result;
-  }
-
-  @Override
-  public VariableSimpleIF getDataVariable(String shortName) {
-    return ncd.getRootGroup().findVariableLocal(shortName);
-  }
-
-  @Override
-  public NetcdfFile getNetcdfFile() {
-    return ncd;
-  }
-
-  private void addGeoGrid(VariableDS varDS, GridCoordSys gcs, Formatter parseInfo) {
-    Gridset gridset;
-    if (null == (gridset = gridsetHash.get(gcs.getName()))) {
-      gridset = new Gridset(gcs);
-      gridsetHash.put(gcs.getName(), gridset);
-      if (parseInfo != null)
-        parseInfo.format(" -make new GridCoordSys= %s%n", gcs.getName());
-      gcs.makeVerticalTransform(this, parseInfo); // delayed until now LOOK why for each grid ??
-    }
-
-    GeoGrid geogrid = new GeoGrid(this, varDS, gridset.gcc);
-    grids.add(geogrid);
-    gridset.add(geogrid);
-  }
-
-  /**
-   * the name of the dataset is the last part of the location
-   * 
-   * @return the name of the dataset
-   */
-  public String getName() {
-    String loc = ncd.getLocation();
-    int pos = loc.lastIndexOf('/');
-    if (pos < 0)
-      pos = loc.lastIndexOf('\\');
-    return (pos < 0) ? loc : loc.substring(pos + 1);
-  }
-
-  /**
-   * @return the underlying NetcdfDataset
-   */
-  public NetcdfDataset getNetcdfDataset() {
-    return ncd;
-  }
-
-  /**
-   * @return the list of GeoGrid objects contained in this dataset.
-   */
-  public List<GridDatatype> getGrids() {
-    return new ArrayList<>(grids);
-  }
-
-  public GridDatatype findGridDatatype(String name) {
-    return findGridByName(name);
-  }
-
-  /**
-   * Return GridDatatype objects grouped by GridCoordSys. All GridDatatype in a Gridset
-   * have the same GridCoordSystem.
-   *
-   * @return List of type ucar.nc2.dt.GridDataset.Gridset
-   */
-  public List<ucar.nc2.dt.GridDataset.Gridset> getGridsets() {
-    return new ArrayList<>(gridsetHash.values());
-  }
-
-  /**
-   * find the named GeoGrid.
-   *
-   * @param fullName find this GeoGrid by full name
-   * @return the named GeoGrid, or null if not found
-   */
-  public GeoGrid findGridByName(String fullName) {
-    for (GeoGrid ggi : grids) {
-      if (fullName.equals(ggi.getFullName()))
-        return ggi;
-    }
-    return null;
-  }
-
-  /**
-   * find the named GeoGrid.
-   *
-   * @param shortName find this GeoGrid by short name
-   * @return the named GeoGrid, or null if not found
-   */
-  public GeoGrid findGridByShortName(String shortName) {
-    for (GeoGrid ggi : grids) {
-      if (shortName.equals(ggi.getShortName()))
-        return ggi;
-    }
-    return null;
-  }
-
-  public GeoGrid findGridDatatypeByAttribute(String attName, String attValue) {
-    for (GeoGrid ggi : grids) {
-      for (Attribute att : ggi.attributes())
-        if (attName.equals(att.getShortName()) && attValue.equals(att.getStringValue()))
-          return ggi;
-    }
-    return null;
-  }
-
-  /**
-   * Get Details about the dataset.
-   */
-  public String getDetailInfo() {
-    Formatter buff = new Formatter();
-    getDetailInfo(buff);
-    return buff.toString();
-  }
-
-  public void getDetailInfo(Formatter buff) {
-    getInfo(buff);
-    buff.format("%n%n----------------------------------------------------%n");
-    buff.format("%s", ncd.toString());
-    buff.format("%n%n----------------------------------------------------%n");
-  }
-
-  /**
-   * Show Grids and coordinate systems.
-   * 
-   * @param buf put info here
-   */
-  private void getInfo(Formatter buf) {
-    int countGridset = 0;
-
-    for (Gridset gs : gridsetHash.values()) {
-      GridCoordSystem gcs = gs.getGeoCoordSystem();
-      buf.format("%nGridset %d  coordSys=%s", countGridset, gcs);
-      buf.format(" LLbb=%s ", gcs.getLatLonBoundingBox());
-      if ((gcs.getProjection() != null) && !gcs.getProjection().isLatLon())
-        buf.format(" bb= %s", gcs.getBoundingBox());
-      buf.format("%n");
-      buf.format("Name__________________________Unit__________________________hasMissing_Description%n");
-      for (GridDatatype grid : gs.getGrids()) {
-        buf.format("%s%n", grid.getInfo());
-      }
-      countGridset++;
-      buf.format("%n");
-    }
-
-    buf.format("%nGeoReferencing Coordinate Axes%n");
-    buf.format("Name__________________________Units_______________Type______Description%n");
-    for (CoordinateAxis axis : ncd.getCoordinateAxes()) {
-      if (axis.getAxisType() == null)
-        continue;
-      axis.getInfo(buf);
-      buf.format("%n");
-    }
-  }
-
-  /**
-   * This is a set of GeoGrids with the same GeoCoordSys.
-   */
-  public static class Gridset implements ucar.nc2.dt.GridDataset.Gridset {
-
-    private final GridCoordSys gcc;
-    private final List<GridDatatype> grids = new ArrayList<>();
-
-    private Gridset(GridCoordSys gcc) {
-      this.gcc = gcc;
-    }
-
-    private void add(GeoGrid grid) {
-      grids.add(grid);
-    }
-
-    /**
-     * Get list of GeoGrid objects
-     */
-    public List<GridDatatype> getGrids() {
-      return grids;
-    }
-
-    /**
-     * all GridDatatype point to this GridCoordSystem
-     */
-    public GridCoordSystem getGeoCoordSystem() {
-      return gcc;
-    }
-
-    /**
-     * all GeoGrids point to this GeoCoordSysImpl.
-     *
-     * @deprecated use getGeoCoordSystem() if possible.
-     */
-    public GridCoordSys getGeoCoordSys() {
-      return gcc;
-    }
-
-  }
-
-  ////////////////////////////
-  // for ucar.nc2.ft.FeatureDataset
-
-  public FeatureType getFeatureType() {
-    return FeatureType.GRID;
-  }
-
-  public String getImplementationName() {
-    return ncd.getConventionUsed();
-  }
-
-  //////////////////////////////////////////////////
-  // FileCacheable
-
-  @Override
-  public synchronized void close() throws java.io.IOException {
-    if (fileCache != null) {
-      if (fileCache.release(this))
-        return;
-    }
-
-    try {
-      if (ncd != null)
-        ncd.close();
-    } finally {
-      ncd = null;
-    }
-  }
-
-  /** @deprecated do not use */
-  @Deprecated
-  public void release() throws IOException {
-    if (ncd != null)
-      ncd.release();
-  }
-
-  /** @deprecated do not use */
-  @Deprecated
-  public void reacquire() throws IOException {
-    if (ncd != null)
-      ncd.reacquire();
-  }
-
-  @Override
-  public long getLastModified() {
-    return (ncd != null) ? ncd.getLastModified() : 0;
-  }
-
-  protected FileCacheIF fileCache;
-
-  /** @deprecated do not use */
-  @Deprecated
-  @Override
-  public synchronized void setFileCache(FileCacheIF fileCache) {
-    this.fileCache = fileCache;
-  }
-
-
-  /////////////////////////////
-  // deprecated
-
-
-  /**
-   * Open a netcdf dataset, parse Conventions, find all the geoGrids, return a GridDataset.
-   *
-   * @deprecated : use GridDataset.open().
-   */
-  public static GridDataset factory(String netcdfFileURI) throws java.io.IOException {
-    return open(netcdfFileURI);
-  }
-
-}
