--- cdm-core/src/main/java/ucar/nc2/dt/grid/package-info.java
@@ -1,20 +1,0 @@
-/*
- * Copyright (c) 1998-2021 John Caron and University Corporation for Atmospheric Research/Unidata
- * See LICENSE for license information.
- */
-
-/**
- * DEPRECATED API - WILL BE REPLACED: Scientific data type "Grid", uses cartesian coordinate system.
- * <p>
- * Deprecated - use ucar.nc2.grid
- * </p>
- * <p>
- * A <b><i>grid</i></b> has
- * a georeferencing &quot;gridded&quot; CoordinateSystem, which implies a connected
- * topology such that values next to each other in index space are next to each
- * other in coordinate space. In the VisAD data model, a grid is a sampled Field.
- * <i> <b>Note: these classes should be considered experimental and may be refactored
- * in the next release. </b></i>
- * </p>
- */
-package ucar.nc2.dt.grid;
